move files and fix small things.
